pyproject.toml
[project]
name = "yamulite"
version = "0.1.0"
description = "Lightweight cross-platform Yandex Music desktop client"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.6",
    "yandex-music>=2.2.0",
    "python-vlc>=3.0.18",
    "requests>=2.31",
    "keyring>=24.0",
]

[project.scripts]
yamulite = "yamulite.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["yamulite*"]
